converting simulation parameter "min num steps" to type long
git-svn-id: file:///Users/sramsey/Desktop/DizzyRepo/ISBJava_Dizzy@2352 5021c4a6-a05a-4b22-8634-7b27b77db61f

+++ ISBJava_Dizzy/trunk/bolouri/java/org/systemsbiology/chem/SimulatorDeterministicRungeKuttaAdaptive.java
@@ -184,11 +184,11 @@ public class SimulatorDeterministicRungeKuttaAdaptive extends SimulatorDetermini
             throw new IllegalArgumentException("max fractional error must be specified");
         }
 
-        int minNumSteps;
-        Integer minNumStepsObj = pSimulatorParams.getMinNumSteps();
+        long minNumSteps;
+        Long minNumStepsObj = pSimulatorParams.getMinNumSteps();
         if(null != minNumStepsObj)
         {
-            minNumSteps = minNumStepsObj.intValue();
+            minNumSteps = minNumStepsObj.longValue();
             if(minNumSteps <= 0)
             {
                 throw new IllegalArgumentException("illegal value for number of steps");
