+++ ISBJava_Dizzy/trunk/bolouri/java/org/systemsbiology/chem/SimulatorDeterministicRungeKuttaFixed.java
@@ -132,11 +132,11 @@ public class SimulatorDeterministicRungeKuttaFixed extends SimulatorDeterministi
             throw new IllegalArgumentException("max relative error must be specified");
         }
 
-        int minNumSteps;
-        Integer minNumStepsObj = pSimulatorParams.getMinNumSteps();
+        long minNumSteps;
+        Long minNumStepsObj = pSimulatorParams.getMinNumSteps();
         if(null != minNumStepsObj)
         {
-            minNumSteps = minNumStepsObj.intValue();
+            minNumSteps = minNumStepsObj.longValue();
             if(minNumSteps <= 0)
             {
                 throw new IllegalArgumentException("illegal value for number of steps");
